feat: auto-merge approved PRs immediately instead of waiting for daemon cycle
Performance improvement: Approved PRs no longer wait up to 60 seconds for the
manager daemon's next check cycle. Auto-merge is triggered immediately after
PR approval, eliminating idle time.

Changes:
- Created src/utils/auto-merge.ts - Shared utility for auto-merging approved PRs
- Updated manager.ts - Use shared auto-merge utility instead of duplicated code
- Updated pr.ts - Call autoMergeApprovedPRs() immediately after approval

Benefits:
- PRs merge immediately when checks pass (0s delay vs up to 60s)
- Reduces time-to-merge for approved work
- Manager daemon still retries as backup if immediate merge fails
- No breaking changes - fully backwards compatible

Fixes STORY-IMP-016
